Moved Playlist Adapter and PlaylistModel to ViewHolders package
